+++ src/Products/PloneMeeting/skins/plonemeeting_styles/plonemeeting.js
@@ -768,9 +768,10 @@ $('table.faceted-table-results').tableDnD({
 });
 }
 
-window.onbeforeunload = function() {
+// do not redefine window.onbeforeunload or it breaks form unload protection
+$(document).ready(function () {
     localStorage.removeItem("toggleAllDetails");
-};
+});
 
 function toggleAllDetails() {
   state = localStorage.getItem("toggleAllDetails");